Adjust Topping SVG styles for ideal output
Toppings reuse donut-topping-buttons asset library.
Every topping added will rotate in increments of 60deg.
Translation offsets for each rotation are too complicated to calculate so instead, styles are mapped to an array and are selected by an index.
A few styles are applied to help clean up SVG assets that target elements tagged with .donut-topping-output.